Sugerencias de tipos y reordenacion de archivos (Usuarios)
- Creado un archivo independiente para cada clase
- Añadido archivo admin.py (No hay nada por ahora)
- Añadida sugerencia de tipos (temporal) para todos los métodos
- Pequeños ajustes en el código inicial
