+++ admin.py
@@ -1,0 +1,10 @@
+from user import User
+
+class Admin(User):
+    @staticmethod
+    def eliminar_usuario(usuario: str) -> None:
+        pass #TODO: Eliminar usuario a partir de su username
+
+    @staticmethod
+    def eliminar_post(nombre_post: str) -> None:
+        pass #TODO: Eliminar post a partir de su nombre
